+++ src/main/java/com/nyancoin/ABCBooks/Book/service/UserService.java
@@ -5,7 +5,7 @@ import com.nyancoin.ABCBooks.Book.domain.User;
 import java.util.List;
 
 public interface UserService {
-    Long add(final User user);
+    int add(final User user);
     void delete(final Long id);
     User get(final Long id);
     List<User> getAll();
